implemented-the-authentication-and-authorization-state-in-the-passenger-baggage-and-airlines-component

+++ src/Components/Airline/AirlineList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useRouteLoaderData } from 'react-router-dom';
 import DataTable from '../../util/DataTable'; // Adjust the path based on your directory structure
 import { Container, Box, Typography, Button, IconButton } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
@@ -8,6 +8,7 @@ import { apiService } from '../../util/apiService';
 
 export default function AirlineList() {
   const [airlines, setAirlines] = useState([]);
+  const auth = useRouteLoaderData('root');
 
   useEffect(() => {
     loadAirlines();
@@ -40,56 +41,59 @@ export default function AirlineList() {
       headerName: 'Name',
       width: 300,
       renderCell: (params) => (
-        <Link to={`/airline-details/${params.row.airlineId}`} style={{ textDecoration: 'none', color: 'inherit' }}>
-          {params.row.name}
-        </Link>
+        auth ? (
+          <Link to={`${params.row.airlineId}`} style={{ textDecoration: 'none', color: 'inherit' }}>
+            {params.row.name}
+          </Link>) : (<span>{params.row.name}</span>)
       )
     },
     { field: 'code', headerName: 'Code', width: 240 },
-    {
-      field: 'edit',
-      headerName: 'Edit',
-      width: 130,
-      renderCell: (params) => (
-        <>
-          <Link to={`/edit-airline/${params.row.airlineId}`} style={{ textDecoration: 'none' }}>
+  ];
+
+  if (auth) {
+    columns.push(
+      {
+        field: 'edit',
+        headerName: 'Edit',
+        width: 100,
+        renderCell: (params) => (
+          <Link to={`${params.row.airlineId}/edit`} style={{ textDecoration: 'none' }}>
             <IconButton color="primary">
               <EditIcon />
             </IconButton>
           </Link>
-        </>
-      )
-    },
-    {
-      field: 'delete',
-      headerName: 'Delete',
-      width: 130,
-      renderCell: (params) => (
-        <>
+        )
+      },
+      {
+        field: 'delete',
+        headerName: 'Delete',
+        width: 100,
+        renderCell: (params) => (
           <IconButton
             color="secondary"
             onClick={() => handleDeleteAirline(params.row.airlineId)}
           >
             <DeleteIcon />
           </IconButton>
-        </>
-      )
-    }
-  ];
+        )
+      }
+    );
+  }
 
   return (
     <Container
-      className='home'
       maxWidth="lg">
       <Box textAlign="center"> {/* mt={5} mb={2} */}
         <Typography variant="h4" component="h1" gutterBottom>
           Airlines
         </Typography>
-        <Link to={`/add-airline`} style={{ textDecoration: 'none' }}>
-          <Button variant="contained" color="primary" style={{ marginRight: '10px' }}>
-            Add Airline
-          </Button>
-        </Link>
+        {auth && (
+          <Link to={`new`} style={{ textDecoration: 'none' }}>
+            <Button variant="contained" color="primary" style={{ marginRight: '10px' }}>
+              Add Airline
+            </Button>
+          </Link>
+        )}
       </Box>
       <Box sx={{ display: 'flex', justifyContent: 'center', marginBottom: 3 }}>
         <Box sx={{ width: '100%', maxWidth: 1000 }}>
